+++ src/moonblade/parser.rs
@@ -41,9 +41,8 @@ impl Rule {
             Self::pow => "pow",
             Self::rem => "mod",
             Self::concat => "concat",
-            // NOTE: and & or operators need to be resolved using if statements
-            // Self::and => "and",
-            // Self::or => "or",
+            Self::and => "and",
+            Self::or => "or",
             Self::not => "not",
             Self::neg => "neg",
 
@@ -345,28 +344,6 @@ fn pratt_parse(pairs: Pairs<Rule>) -> Result<Expr, String> {
                         }
                     }
                 },
-
-                // Short-circuiting and
-                Rule::and => {
-                    let lhs_res = lhs?;
-
-                    // a && b => if(a, b, a)
-                    Expr::Func(FunctionCall::new(
-                        "if",
-                        vec![lhs_res.clone(), rhs?, lhs_res],
-                    ))
-                }
-
-                // Short-circuiting or
-                Rule::or => {
-                    let lhs_res = lhs?;
-
-                    // a || b => if(a, a, b)
-                    Expr::Func(FunctionCall::new(
-                        "if",
-                        vec![lhs_res.clone(), lhs_res, rhs?],
-                    ))
-                }
 
                 // Access & call
                 Rule::point => match rhs? {